+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 0ccdd7a - v16
+| Author: fjrodafo - Wed, 1 Jan 2025 14:09:37 +0000
+| 
 * Commit: 7eed115 - v15
 | Author: fjrodafo - Mon, 30 Dec 2024 14:20:28 +0000
 |